functions/requirements.txt
Flask
flask-cors
Pillow
torch
transformers
pydantic_settings
spacy
firebase-functions